+++ client/src/main/java/replicant/Transport.java
@@ -135,24 +135,14 @@ public interface Transport
 
   void requestSubscribe( @Nonnull ChannelAddress address,
                          @Nullable Object filter,
-                         @Nonnull String eTag,
-                         @Nonnull SafeProcedure onCacheValid,
-                         @Nonnull SafeProcedure onSuccess,
-                         @Nonnull Consumer<Throwable> onError );
-
-  void requestSubscribe( @Nonnull ChannelAddress address,
-                         @Nullable Object filter,
+                         @Nullable String eTag,
+                         @Nullable SafeProcedure onCacheValid,
                          @Nonnull SafeProcedure onSuccess,
                          @Nonnull Consumer<Throwable> onError );
 
   void requestUnsubscribe( @Nonnull ChannelAddress address,
                            @Nonnull SafeProcedure onSuccess,
                            @Nonnull Consumer<Throwable> onError );
-
-  void requestSubscriptionUpdate( @Nonnull ChannelAddress address,
-                                  @Nonnull Object filter,
-                                  @Nonnull SafeProcedure onSuccess,
-                                  @Nonnull Consumer<Throwable> onError );
 
   void requestBulkSubscribe( @Nonnull List<ChannelAddress> addresses,
                              @Nullable Object filter,
@@ -162,11 +152,6 @@ public interface Transport
   void requestBulkUnsubscribe( @Nonnull List<ChannelAddress> addresses,
                                @Nonnull SafeProcedure onSuccess,
                                @Nonnull Consumer<Throwable> onError );
-
-  void requestBulkSubscriptionUpdate( @Nonnull List<ChannelAddress> addresses,
-                                      @Nonnull Object filter,
-                                      @Nonnull SafeProcedure onSuccess,
-                                      @Nonnull Consumer<Throwable> onError );
 
   /**
    * Notify the Transport when a Connector has completed processing a message.